requirements.txt
argon2-cffi==20.1.0
asgiref==3.2.10
cffi==1.14.2
Django==3.1.1
django-cors-headers==3.5.0
django-templated-mail==1.1.1
djangorestframework==3.11.1
djangorestframework-simplejwt==4.4.0
djoser==2.0.5
pycparser==2.20
PyJWT==1.7.1
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
